API documentation (#3)
* Creating simple documentation on OpenAPI format

On this change I'm implementing the generation of API documentation
on OpenAPI format.

At this moment I'm only implementing the basic structure to generate the
documentation, only with the routes and response codes. To the next
steps I'll implement the other stuffs (like models).

I'm creating a command to generate a JSON file with the documentation.
To use it, you only need to run `php bin/console app:generate-doc` on
API root.

* Creating a container to show API documentation

On this change I'm creating a container to show the API documentation
on OpenAPI format.